+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/QueryPerfTest/FeedResponseObjectTags.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using Newtonsoft.Json;
+
+internal sealed class Tags
+{
+    [JsonProperty(PropertyName = "words")]
+    public List<string> Words { get; set; }
+    [JsonProperty(PropertyName = "numbers")]
+    public string Numbers { get; set; }
+}
