+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/QueryPerfTest/QueryStatisticsAccumulator.cs
@@ -133,6 +133,9 @@
                     TransitTime = this.queryStatisticsAccumulatorBuilder.queryMetrics.TransitTime,
                     Received = this.queryStatisticsAccumulatorBuilder.queryMetrics.Received,
                     Completed = this.queryStatisticsAccumulatorBuilder.queryMetrics.Completed,
+                    PocoTimeList = this.queryStatisticsAccumulatorBuilder.queryMetrics.PocoTimeList,
+                    GetCosmosElementResponseTimeList = this.queryStatisticsAccumulatorBuilder.queryMetrics.GetCosmosElementResponseTimeList,
+                    EndToEndTimeList = this.queryStatisticsAccumulatorBuilder.queryMetrics.EndToEndTimeList,
                     BadRequestCreated = this.queryStatisticsAccumulatorBuilder.queryMetrics.BadRequestCreated,
                     BadRequestChannelAcquisitionStarted = this.queryStatisticsAccumulatorBuilder.queryMetrics.BadRequestChannelAcquisitionStarted,
                     BadRequestPipelined = this.queryStatisticsAccumulatorBuilder.queryMetrics.BadRequestPipelined,
