+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/QueryPerfTest/QueryStatisticsAccumulatorBuilder.cs
@@ -1,0 +1,7 @@
+﻿using System.Collections.Generic;
+
+internal class QueryStatisticsAccumulatorBuilder
+{
+    public QueryMetrics queryMetrics = new();
+    public List<QueryMetrics> QueryMetricsList { get; set; } = new List<QueryMetrics>();
+}
