+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/QueryPerfTest/TextWriterExtension.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.IO;
+
+internal static class TextWriterExtension
+{
+    public static void WriteLines(this TextWriter textWriter, params double[] metrics)
+    {
+        foreach (double metric in metrics)
+        {
+            textWriter.WriteLine(metric);
+        }
+    }
+}
